# Analytical 2D and 3D eigenvalues and determinant