--- a/skimage/filters/ridges.py
+++ b/skimage/filters/ridges.py
@@ -15,7 +15,8 @@ import numpy as np
 from scipy import linalg
 
 from .._shared.utils import _supported_float_type, check_nD, deprecated
-from ..feature.corner import hessian_matrix, hessian_matrix_eigvals
+from ..feature.corner import (_symmetric_compute_eigenvalues, hessian_matrix,
+                              hessian_matrix_eigvals)
 from ..util import img_as_float
 
 
@@ -327,10 +328,13 @@ def frangi(image, sigmas=range(1, 10, 2), scale_range=None,
     # from different (sigma) scales
     filtered_max = np.zeros_like(image)
     for sigma in sigmas:  # Filter for all sigmas.
-        eigvals = hessian_matrix_eigvals(hessian_matrix(
-            image, sigma, mode=mode, cval=cval, use_gaussian_derivatives=True))
-        # Sort eigenvalues by magnitude.
-        eigvals = np.take_along_axis(eigvals, abs(eigvals).argsort(0), 0)
+        H = hessian_matrix(image, sigma, mode=mode, cval=cval,
+                           use_gaussian_derivatives=True)
+        # Use _symmetric_compute_eigenvalues rather than
+        # hessian_matrix_eigvals so we can directly sort by ascending magnitude
+        eigvals = _symmetric_compute_eigenvalues(
+            H, sort='ascending', abs_sort=True
+        )
         lambda1 = eigvals[0]
         if image.ndim == 2:
             lambda2, = np.maximum(eigvals[1:], 1e-10)