closes #6635.  Should also give further improvement for issue #6507.

Faster and more memory efficient eigenvalue and determinant computations as used by several feature and filter functions. This is a port of downstream PR rapidsai/cucim#434. Details of the functions benefiting from this are listed in #6635

There are no changes to the public API. Behavior is the same aside from possible numeric precision / floating point rounding differences. New test functions were added to verify a close match to the prior implementation.

The private `_symmetric_compute_eigenvalues` helper gets a couple of optional kwargs to control how the eigenvalues are sorted. Values other than the default are currently only used to avoid additional sorting steps for the `frangi` ridge filter.

The old implementation involved formation of a large `_symmetric_image` intermediate array of shape `(rows, cols, 2, 2)` in 2D or `(planes, rows, cols, 3, 3)` in 3D and then passes that to a `numpy.linalg` function. The new implementation does not form this intermediate array at all, so the large memory overhead is avoided. Using an analytical solution rather than a generic solver is also more performant.

Benchmark results for `hessian_matrix_det` and `hessian_matrix_eigvals` are shown below. The reference `peakmem` is mainly due to computation of the hessian matrix itself. It can be seen that the old implementation substantially increased peak memory usage above that for just computing the hessian itself, while the new one does not. Performance is also substantially better than previously (30-40x for 2D eigenvalues and ~10x for 3D eigenvalues). It is also 4-5x better for the determinant even though the time for this function also involves computation of the hessian matrix itself.

### hessian_matrix_det

shape           | duration (old) | duration (new) | peakmem (old) | peakmem (new) | peakmem (reference)
----------------|----------------|----------------|---------------|---------------|--------------------
(512, 512)      | 0.0555 s       | 0.0100 s       | 98.9 M        | 86.3 M        | 86.3 M
(3840, 2160)    | 1.4400 s       | 0.2530 s       | 837 M         | 528 M         | 528 M
(64, 48, 32)    | 0.0298 s       | 0.0073 s       | 91.7 M        | 81.2 M        | 80.7 M
(128, 96, 64)   | 0.2300 s       | 0.0452 s       | 216 M         | 132 M         | 131 M
(256, 192, 128) | 1.8000 s       | 0.3510 s       | 1210 M        | 528 M         | 528 M

### hessian_matrix_eigvals

shape           | duration (old) | duration (new) | peakmem (old) | peakmem (new) | peakmem (reference)
----------------|----------------|----------------|---------------|---------------|--------------------
(512, 512)      | 0.0650 s       | 0.0018 s       | 101 M         | 83.1 M        | 83.1 M
(3840, 2160)    | 1.8500 s       | 0.0457 s       | 772 M         | 340 M         | 340 M
(64, 48, 32)    | 0.1170 s       | 0.0177 s       | 93 M          | 78.7 M        | 78.7
(128, 96, 64)   | 0.6280 s       | 0.0593 s       | 208 M         | 113 M         | 113 M
(256, 192, 128) | 4.5400 s       | 0.4680 s       | 1110 M        | 376 M         | 376 M