@anntzer 2025-07-06
I would be happy if analytical eigendecomposition of symmetric 3x3 matrices was available; is this PR blocked by anything specific?